refactor(tool_router): isolate deprecated endpoints and update docs
- Move 4 deprecated backward-compat endpoints (GET /kwic/{search},
  GET /word_trend_speeches/{search}, GET+POST /speeches,
  POST /speech_download/) to a new deprecated_endpoints.py router;
  all decorated with deprecated=True and tagged 'Tools (deprecated)'
- Register deprecated_endpoints.router in app.py alongside the active
  tool_router and metadata_router
- Update docs/DESIGN.md: reflect three-router layout and link to the
  new file for easy future removal

Frontend sweep: getKwicResultLegacy, getWordTrendsSpeeches, and
getSpeechesResult are dead code (never called). nGramDataStore still
calls GET /speeches?speech_id=... for ngram speech drill-down; the
deprecated endpoint is retained until that feature migrates.
