+++ api_swedeb/app.py
@@ -10,7 +10,7 @@ from starlette.types import Scope
 
 from api_swedeb.api.container import AppContainer
 from api_swedeb.api.services.result_store import ResultStore
-from api_swedeb.api.v1.endpoints import metadata_router, tool_router
+from api_swedeb.api.v1.endpoints import deprecated_endpoints, metadata_router, tool_router
 from api_swedeb.core.configuration import ConfigValue, get_config_store
 
 # pylint: disable=import-outside-toplevel
@@ -80,6 +80,7 @@ def create_app(*, config_source: str | None = DEFAULT_CONFIG_SOURCE, static_dir:
     )
 
     app.include_router(tool_router.router)
+    app.include_router(deprecated_endpoints.router)
     app.include_router(metadata_router.router)
 
     @app.get("/public/index.html", include_in_schema=False)
